backend/requirements.txt
# Flask Framework
flask==3.0.0
flask-cors==4.0.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Environment Variables
python-dotenv==1.0.0

# Security & Authentication
PyJWT==2.8.0
cryptography==41.0.7

# Utilities
python-dateutil==2.8.2
requests==2.31.0
